Refactor harmonic comb trigger into dedicated module
